Rename some files to fix tests (#26396)
