+++ api/components/com_banners/View/Banners/JsonapiView.php
@@ -1,24 +1,24 @@
 <?php
 /**
  * @package     Joomla.API
- * @subpackage  com_tags
+ * @subpackage  com_banners
  *
  * @copyright   Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
-namespace Joomla\Component\Tags\Api\View\Tags;
+namespace Joomla\Component\Banners\Api\View\Banners;
 
 defined('_JEXEC') or die;
 
 use Joomla\CMS\MVC\View\JsonApiView as BaseApiView;
 
 /**
- * The tags view
+ * The banners view
  *
  * @since  4.0.0
  */
-class JsonApiView extends BaseApiView
+class JsonapiView extends BaseApiView
 {
 	/**
 	 * The fields to render item in the documents
@@ -27,37 +27,42 @@ class JsonApiView extends BaseApiView
 	 * @since  4.0.0
 	 */
 	protected $fieldsToRenderItem = [
-		'id',
-		'parent_id',
-		'level',
-		'lft',
-		'rgt',
-		'alias',
 		'typeAlias',
-		'path',
-		'title',
-		'note',
+		'id',
+		'cid',
+		'type',
+		'name',
+		'alias',
+		'imptotal',
+		'impmade',
+		'clicks',
+		'clickurl',
+		'state',
+		'catid',
 		'description',
-		'published',
+		'custombannercode',
+		'sticky',
+		'ordering',
+		'metakey',
+		'params',
+		'own_prefix',
+		'metakey_prefix',
+		'purchase_type',
+		'track_clicks',
+		'track_impressions',
 		'checked_out',
 		'checked_out_time',
-		'access',
-		'params',
-		'metadesc',
-		'metakey',
-		'metadata',
-		'created_user_id',
-		'created_time',
-		'created_by_alias',
-		'modified_user_id',
-		'modified_time',
-		'images',
-		'urls',
-		'hits',
-		'language',
-		'version',
 		'publish_up',
 		'publish_down',
+		'reset',
+		'created',
+		'language',
+		'created_by',
+		'created_by_alias',
+		'modified',
+		'modified_by',
+		'version',
+		'contenthistoryHelper',
 	];
 
 	/**
@@ -68,25 +73,26 @@ class JsonApiView extends BaseApiView
 	 */
 	protected $fieldsToRenderList = [
 		'id',
-		'title',
+		'name',
 		'alias',
-		'note',
-		'published',
-		'access',
-		'description',
 		'checked_out',
 		'checked_out_time',
-		'created_user_id',
-		'path',
-		'parent_id',
-		'level',
-		'lft',
-		'rgt',
+		'catid',
+		'clicks',
+		'metakey',
+		'sticky',
+		'impmade',
+		'imptotal',
+		'state',
+		'ordering',
+		'purchase_type',
 		'language',
-		'language_title',
+		'publish_up',
+		'publish_down',
 		'language_image',
 		'editor',
-		'author_name',
-		'access_title',
+		'category_title',
+		'client_name',
+		'client_purchase_type',
 	];
 }
