+++ api/components/com_config/Controller/ApplicationController.php
@@ -14,7 +14,7 @@ defined('_JEXEC') or die;
 use Joomla\CMS\Access\Exception\NotAllowed;
 use Joomla\CMS\Form\Form;
 use Joomla\CMS\MVC\Controller\ApiController;
-use Joomla\Component\Config\Api\View\Application\JsonApiView;
+use Joomla\Component\Config\Api\View\Application\JsonapiView;
 use Joomla\Component\Config\Administrator\Model\ApplicationModel;
 use Tobscure\JsonApi\Exception\InvalidParameterException;
 
@@ -55,7 +55,7 @@ class ApplicationController extends ApiController
 
 		try
 		{
-			/** @var JsonApiView $view */
+			/** @var JsonapiView $view */
 			$view = $this->getView(
 				$this->default_view,
 				$viewType,
