+++ api/components/com_fields/View/Fields/JsonapiView.php
@@ -19,7 +19,7 @@ use Joomla\CMS\Router\Exception\RouteNotFoundException;
  *
  * @since  4.0.0
  */
-class JsonApiView extends BaseApiView
+class JsonapiView extends BaseApiView
 {
 	/**
 	 * The fields to render item in the documents
