+++ api/components/com_languages/View/Strings/JsonapiView.php
@@ -21,7 +21,7 @@ use Tobscure\JsonApi\Collection;
  *
  * @since  4.0.0
  */
-class JsonApiView extends BaseApiView
+class JsonapiView extends BaseApiView
 {
 	/**
 	 * The fields to render items in the documents
