--- api/components/com_menus/View/Menus/JsonApiView.php
@@ -1,54 +1,0 @@
-<?php
-/**
- * @package     Joomla.API
- * @subpackage  com_menus
- *
- * @copyright   Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
- * @license     GNU General Public License version 2 or later; see LICENSE.txt
- */
-
-namespace Joomla\Component\Menus\Api\View\Menus;
-
-defined('_JEXEC') or die;
-
-use Joomla\CMS\MVC\View\JsonApiView as BaseApiView;
-
-/**
- * The menus view
- *
- * @since  4.0.0
- */
-class JsonApiView extends BaseApiView
-{
-	/**
-	 * The fields to render item in the documents
-	 *
-	 * @var  array
-	 * @since  4.0.0
-	 */
-	protected $fieldsToRenderItem = [
-		'id',
-		'menutype',
-		'title',
-		'description',
-		'client_id',
-		'count_published',
-		'count_unpublished',
-		'count_trashed',
-	];
-
-	/**
-	 * The fields to render items in the documents
-	 *
-	 * @var  array
-	 * @since  4.0.0
-	 */
-	protected $fieldsToRenderList  = [
-		'id',
-		'asset_id',
-		'menutype',
-		'title',
-		'description',
-		'client_id',
-	];
-}
