+++ api/components/com_privacy/View/Requests/JsonapiView.php
@@ -23,7 +23,7 @@ use Tobscure\JsonApi\Resource;
  *
  * @since  4.0.0
  */
-class JsonApiView extends BaseApiView
+class JsonapiView extends BaseApiView
 {
 	/**
 	 * The fields to render item in the documents
